backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Hannes221 <hannespfau@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
ansible = "^9.3.0"
uvicorn = "^0.29.0"
predictionguard = "^1.9.2"
redis = "^5.0.3"
sqlalchemy = "^2.0.28"
pydantic-settings = "^2.2.1"
psycopg2 = "^2.9.9"
uuid = "^1.30"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
